+++ src/endpoints/basket/getBasket.ts
@@ -1,91 +1,12 @@
 import {prefixList} from '../../helpers/attributes';
 import {BapiCall} from '../../helpers/execute';
-import {BapiPrice, Product, Variant} from '../../types/Product';
+import {BasketResponse} from '../../types/Basket';
 import {
   ProductWith,
   productWithQueryParameterValues,
   VariantWith,
   variantWithQueryParameterValues,
 } from '../../types/ProductWith';
-
-export type BasketItemPrice = Omit<BapiPrice, 'tax'>;
-
-export type BasketTotalPrice = Omit<BapiPrice, 'tax' | 'reference'>;
-
-export type BasketKey = string & {
-  readonly ____tagBasketKey: unique symbol;
-};
-
-export interface BasketResponseData<P = Product, V = Variant> {
-  key: BasketKey;
-  cost: BasketTotalPrice;
-  currencyCode: 'EUR';
-  items: BasketItem<P, V>[];
-
-  packages: BasketPackageInformation[];
-}
-
-export interface ItemGroup {
-  id: string;
-  isMainItem: boolean;
-  isRequired: boolean;
-}
-
-export interface BasketItem<P = Product, V = Variant> {
-  key: string;
-  customData: unknown;
-  packageId: number;
-  price: {
-    total: BasketItemPrice;
-    unit: BasketItemPrice;
-  };
-  quantity: number;
-  availableQuantity?: number;
-  deliveryForecast?: {
-    deliverable?: {
-      quantity: number;
-      key: string;
-    };
-    subsequentDelivery?: {
-      quantity: number;
-      key: string;
-    };
-  };
-  status: 'available' | 'unavailable' | 'deleted';
-  product: P;
-  variant: V;
-  displayData: BasketItemDisplayData;
-  itemGroup?: ItemGroup;
-}
-
-export interface BaskteItemDisplayDataItem {
-  value: string;
-  label: string;
-  key: string;
-}
-
-export type BasketItemDisplayDataKey =
-  | 'meta'
-  | 'name'
-  | 'identifier'
-  | 'attribute-1'
-  | 'attribute-2'
-  | 'attribute-3';
-
-export type BasketItemDisplayData = Partial<
-  Record<BasketItemDisplayDataKey, BaskteItemDisplayDataItem>
->;
-
-export interface BasketPackageInformation {
-  id: number;
-  carrierKey: string;
-  deliveryDate: {
-    // Date like '2018-02-05'
-    max: string;
-    // Date like '2018-02-02'
-    min: string;
-  };
-}
 
 export interface BasketWith {
   items?: {
@@ -129,7 +50,7 @@ export interface GetBasketParameters {
 
 export function getBasketEndpointRequest(
   params: GetBasketParameters,
-): BapiCall<BasketResponseData> {
+): BapiCall<BasketResponse> {
   return {
     method: 'GET',
     endpoint: `/v1/baskets/${params.basketKey}`,
