+++ src/endpoints/shopconfiguration/shopconfiguration.ts
@@ -1,15 +1,7 @@
 import {BapiCall} from '../../helpers/execute';
+import {ShopConfigurationResponse} from '../../types/ShopConfiguration';
 
-export interface ShopConfigurationResponseData {
-  shopId: number;
-  name: string;
-  properties: Array<{
-    key: string;
-    value: string;
-  }>;
-}
-
-export function createShopConfigurationRequest(): BapiCall<ShopConfigurationResponseData> {
+export function createShopConfigurationRequest(): BapiCall<ShopConfigurationResponse> {
   return {
     method: 'GET',
     endpoint: '/v1/shop-configuration',
